+++ codeveloper_tools/windows/codeveloper_tools_plugin.cpp
@@ -1,0 +1,9 @@
+#include "include/codeveloper_tools/codeveloper_tools_plugin.h"
+
+void CodeveloperToolsPluginRegisterWithRegistrar(
+    FlutterDesktopPluginRegistrarRef registrar) {
+  // Register the plugin with the Flutter engine.
+  CodeveloperToolsPlugin::RegisterWithRegistrar(
+      flutter::PluginRegistrarManager::GetInstance()
+          ->GetRegistrar<flutter::PluginRegistrarWindows>(registrar));
+}